+++ pkg/chunk/storage/factory.go
@@ -27,7 +27,6 @@ import (
 	"github.com/grafana/mimir/pkg/chunk/local"
 	"github.com/grafana/mimir/pkg/chunk/objectclient"
 	"github.com/grafana/mimir/pkg/chunk/openstack"
-	"github.com/grafana/mimir/pkg/chunk/purger"
 	util_log "github.com/grafana/mimir/pkg/util/log"
 )
 
@@ -98,8 +97,6 @@ type Config struct {
 
 	IndexQueriesCacheConfig cache.Config `yaml:"index_queries_cache_config"`
 
-	DeleteStoreConfig purger.DeleteStoreConfig `yaml:"delete_store"`
-
 	GrpcConfig grpc.Config `yaml:"grpc_store"`
 }
 
@@ -112,7 +109,6 @@ func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
 	cfg.CassandraStorageConfig.RegisterFlags(f)
 	cfg.BoltDBConfig.RegisterFlags(f)
 	cfg.FSConfig.RegisterFlags(f)
-	cfg.DeleteStoreConfig.RegisterFlags(f)
 	cfg.Swift.RegisterFlags(f)
 	cfg.GrpcConfig.RegisterFlags(f)